pyproject.toml
[tool.poetry]
name = "hexlet_code"
version = "0.5.0"
description = ""
authors = ['Roddy <per677199@yandex.ru>']

packages = [
{ include = 'gendiff.py'},
{ include = 'diff_generator'}
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest = "^6.2.1"
pip = "^20.3.3"
coverage = "^5.3.1"
codeclimate-test-reporter = "^0.2.3"
pytest-cov = "^2.10.1"

[tool.poetry.scripts]
gendiff = "gendiff:generate_diff"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

